+++ CHANGELOG.md
@@ -27,6 +27,10 @@ All notable changes to this project will be documented in this file.
 ### Added
 - welcome post
 
+### Fixed
+- social links
+- row bug in Mirko
+
 ## 0.3.7 - 2017-02-12
 
 ### Fixed